% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bom_response.R
\name{body_error}
\alias{body_error}
\title{Check for errors in the body of a BOM API response}
\usage{
body_error(resp)
}
\arguments{
\item{resp}{A response object returned by the BOM API}
}
\value{
An error message
}
\description{
Error messages are often contained in the body of the response and are not
picked up by \code{httr} by default. This function extracts the error message in a
way appropriate to each content type (json, xml, or html). Note that if an
error occurs the returned content type does not necessarily match the
requested format.
}
