% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bom_response.R
\name{get_bom_response}
\alias{get_bom_response}
\title{Get a response from the BOM API}
\usage{
get_bom_response(format, request, ...)
}
\arguments{
\item{format}{A string giving the format of the response content. Valid
options will depend on the request. Use \code{\link[=list_formats]{list_formats()}} to see available
options.}

\item{request}{A character string defining the request type to pass on to the
BOM API. Use \code{\link[=list_requests]{list_requests()}} to see available options.}

\item{...}{Optional named query fields which can be used to narrow the
request. Vectors of length greater than one will be collapsed into a comma
separated character string. Use \code{\link[=list_query_fields]{list_query_fields()}} to see available
options, including fields which accept a comma separated list or wildcard.}
}
\value{
A response object.
}
\description{
\code{get_bom_response} takes an arbitrary API request and returns the response
object. Inputs are not checked in any way, allowing you to make requests
which fall outside the scope of \code{\link[=get_bom_data]{get_bom_data()}} or \code{\link[=get_timeseries]{get_timeseries()}}. This
can be useful if you want to return an unconventional format (e.g. geojson),
or just want to interface directly with the API.
}
\examples{
# Get a geojson file of discharge stations
get_bom_response(
  format = "geojson",
  request = "getStationList",
  parametertype_name = "Water Course Discharge"
)
}
\seealso{
\code{\link[=get_bom_data]{get_bom_data()}} for a simplified API interface which is suited to
requests for rectangular data.
}
