% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{get_timeseries}
\alias{get_timeseries}
\title{Download a timeseries for a specific station}
\usage{
get_timeseries(
  station_no = NULL,
  ts_name = NULL,
  parametertype_name = NULL,
  from = NULL,
  to = NULL,
  timezone = "UTC",
  ...,
  returnfields = c("Timestamp", "Value", "Quality Code", "Interpolation Type")
)
}
\arguments{
\item{station_no}{Optionally, a numeric or character vector of station
numbers.}

\item{ts_name}{Optionally, the name of a specific time series}

\item{parametertype_name}{Optionally, a character vector of parameters. This
can be useful if, for instance, you are only interested in stations which
measure "Water Course Discharge". Use \code{\link[=list_parameters]{list_parameters()}} to see available
options.}

\item{from, to}{Optionally, timestamps defining the the start and end points
of the requested timeseries. \code{from} and \code{to} default to the begining and
end of the time series respectively. If no time is specified (e.g.
YYYY-MM-DD format) then values will be returned for the whole day.}

\item{timezone}{The timezone in which to return time stamps. Defaults to UTC.
Note that stations record measurements in non-daylight savings local time.}

\item{...}{Optional named query fields which can be used to narrow the
request. Vectors of length greater than one will be collapsed into a comma
separated character string. Use \code{\link[=list_query_fields]{list_query_fields()}} to see available
options, including fields which accept a comma separated list or wildcard.}

\item{returnfields}{A character vector of columns to include in the returned
tibble. If not defined, the default columns will be determined by the API.
Use \code{\link[=list_return_fields]{list_return_fields()}} to see available options.}
}
\value{
A tibble with columns matching \code{returnfields}.
}
\description{
Download a timeseries for a specific station
}
\examples{
get_timeseries(
  station_no = c("425004", "423005"),
  from = "2020-01-01",
  to = "2020-01-31",
  ts_name = "DMQaQc.Merged.AsStored.1",
  parametertype_name = "Water Course Level"
)
}
