% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{get_timeseries_values}
\alias{get_timeseries_values}
\title{Download a specific timeseries}
\usage{
get_timeseries_values(
  ts_id,
  from = NULL,
  to = NULL,
  timezone = "UTC",
  ...,
  md_returnfields = c("ts_id", "station_no"),
  returnfields = c("Timestamp", "Value", "Quality Code", "Interpolation Type")
)
}
\arguments{
\item{ts_id}{One or more timeseries ids. \code{ts_id} specifies both the time
series product and the station.}

\item{from, to}{Optionally, timestamps defining the the start and end points
of the requested timeseries. \code{from} and \code{to} default to the begining and
end of the time series respectively. If no time is specified (e.g.
YYYY-MM-DD format) then values will be returned for the whole day.}

\item{timezone}{The timezone in which to return time stamps. Defaults to UTC.
Note that stations record measurements in non-daylight savings local time.}

\item{...}{Optional named query fields which can be used to narrow the
request. Vectors of length greater than one will be collapsed into a comma
separated character string. Use \code{\link[=list_query_fields]{list_query_fields()}} to see available
options, including fields which accept a comma separated list or wildcard.}

\item{md_returnfields}{additional metadata fields to include in the response}

\item{returnfields}{A character vector of columns to include in the returned
tibble. If not defined, the default columns will be determined by the API.
Use \code{\link[=list_return_fields]{list_return_fields()}} to see available options.}
}
\value{
A tibble with columns matching \code{returnfields}.
}
\description{
Download a specific timeseries
}
\examples{
get_timeseries_values(ts_id = 83527010, from = "2021-01-01", to = "2021-01-07")
}
