% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response.R
\name{get_bom_data}
\alias{get_bom_data}
\title{Get BOM data}
\usage{
get_bom_data(request, ...)
}
\arguments{
\item{request}{A character string defining the request type to pass on to the
KISTERS API. Use \code{\link[=list_requests]{list_requests()}} to see available requests.}

\item{...}{Optional key value pairs passed on to \code{\link[=get_bom_response]{get_bom_response()}}.}
}
\value{
A tibble.
}
\description{
Get BOM data
}
\examples{
get_bom_data(request = "getStationList")
}
\seealso{
\code{\link[=get_bom_response]{get_bom_response()}} for more fine-grained control over API
requests.
}
