% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response.R
\name{get_bom_data}
\alias{get_bom_data}
\title{Get BOM data}
\usage{
get_bom_data(request, ...)
}
\arguments{
\item{request}{A character string defining the request type to pass on to the
BOM API. Use \code{\link[=list_requests]{list_requests()}} to see available options.}

\item{...}{Optional named query fields which can be used to narrow the
request. Vectors of length greater than one will be collapsed into a comma
separated list. It is possible to use an asterix character (*) as wildcard
in some fields, however this is not recommended as the naming conventions
can be inconsistent. Use \code{\link[=list_query_fields]{list_query_fields()}} to see available options,
including fields which accept a comma separated list or wildcard.}
}
\value{
A tibble.
}
\description{
Get BOM data
}
\examples{
get_bom_data(request = "getStationList")
}
\seealso{
\code{\link[=get_bom_response]{get_bom_response()}} for more fine-grained control over API
requests.
}
