% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response.R
\name{get_bom_data}
\alias{get_bom_data}
\title{Get BOM data}
\usage{
get_bom_data(request, ...)
}
\arguments{
\item{request}{Desired request e.g. "getStationList"}

\item{...}{Optional key value pairs, passed on to \code{get_bom_response}}
}
\value{
A tibble
}
\description{
Get BOM data
}
\examples{
get_bom_data(request = "getStationList")
}
