% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_funcs.R
\name{get_bom_data}
\alias{get_bom_data}
\title{Get BOM data}
\usage{
get_bom_data(request, ..., returnfields = NULL)
}
\arguments{
\item{request}{A character string defining the request type to pass on to the
BOM API. Use \code{\link[=list_requests]{list_requests()}} to see available options.}

\item{...}{Optional named query fields which can be used to narrow the
request. Vectors of length greater than one will be collapsed into a comma
separated list. Use \code{\link[=list_query_fields]{list_query_fields()}} to see available options,
including fields which accept a comma separated list or wildcard.}

\item{returnfields}{A character vector of columns to include in the returned
tibble. If not defined, the default columns will be determined by the API.
Use \code{\link[=list_return_fields]{list_return_fields()}} to see available options.}
}
\value{
A tibble with columns determined by \code{returnfields}.
}
\description{
Get BOM data
}
\examples{
get_bom_data(request = "getStationList")
}
\seealso{
\code{\link[=get_bom_response]{get_bom_response()}} for more fine-grained control over API
requests.
}
