% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response.R
\name{get_bom_response}
\alias{get_bom_response}
\title{Get a response from the BOM API}
\usage{
get_bom_response(format, request, ..., max_tries = 1)
}
\arguments{
\item{format}{Response content format}

\item{request}{A character string defining the request type to pass on to the
KISTERS API. Use \code{\link[=list_requests]{list_requests()}} to see available requests.}

\item{...}{Optional key value pairs which can be used to narrow down the
request. Vectors of length greater than one will be collapsed into a comma
separated list. It is possible to use an asterix character (*) as wildcard
in some fields, however this is not recommended as the data naming
conventions can be inconsistent. Use \code{\link[=list_query_fields]{list_query_fields()}} to see available
options.}

\item{max_tries}{Optional. Maximum number of times to retry the request if an
error is returned. Passed on to \code{httr2::req_retry}.}
}
\value{
A response object.
}
\description{
Make an arbitrary request to the API, and receive the response object. This
function is intentionally permissive, meaning that it will allow you to make
any request that you like (so long as you specify format and request). No
error checking is perfomed on the inputs.
}
\examples{
# Get a geojson file of discharge stations
get_bom_response(
  format = "geojson",
  request = "getStationList",
  parametertype_name = "Water Course Discharge"
)
}
\seealso{
\code{\link[=get_bom_data]{get_bom_data()}} for a simplified API interface which is suitable
for most requests returning rectangular data.
}
