% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response.R
\name{get_bom_response}
\alias{get_bom_response}
\title{Get a response from the BOM API}
\usage{
get_bom_response(format, request, ..., max_tries = 1)
}
\arguments{
\item{format}{Response content format}

\item{request}{Request type}

\item{...}{Optional key value pairs}

\item{max_tries}{Maximum number of times to retry the request if an error is
returned}
}
\value{
A response object
}
\description{
Get a response from the BOM API
}
\examples{
get_bom_response(
  format = "json",
  request = "getStationList",
  parametertype_name = "Water Course Discharge"
)
}
