% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_info.R
\name{get_station_list}
\alias{get_station_list}
\title{Get a list of gauging stations}
\usage{
get_station_list(
  station_no = NULL,
  station_name = NULL,
  parametertype_name = NULL,
  bbox = NULL,
  ...,
  returnfields = c("station_no", "station_name", "station_latitude", "station_longitude")
)
}
\arguments{
\item{station_no}{Optionally, a character vector of station numbers.}

\item{station_name}{Optionally, a character vector of station names.}

\item{parametertype_name}{Optionally, a character vector of parameters. This
can be useful if, for instance, you are only interested in stations which
measure 'Water Course Discharge'. Use \code{\link[=list_parameters]{list_parameters()}} to see available
options.}

\item{bbox}{Optionally, a numeric vector of the form \code{c(min_x, min_y, max_x, max_y)} which can be used to get stations in a specific bounding box.}

\item{...}{Optional named query fields which can be used to narrow the
request. Vectors of length greater than one will be collapsed into a comma
separated character string. Use \code{\link[=list_query_fields]{list_query_fields()}} to see available
options, including fields which accept a comma separated list or wildcard.}

\item{returnfields}{A character vector of columns to include in the returned
tibble. If not defined, the default columns will be determined by the API.
Use \code{\link[=list_return_fields]{list_return_fields()}} to see available options.}
}
\value{
A tibble with columns matching \code{returnfields}.
}
\description{
\code{get_station_list} queries the BOM API and returns a table of station
details. The list of stations can be narrowed by station number, station
name, parameter, or geographic area (bbox).
}
\examples{
get_station_list()
}
