% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{get_timeseries_list}
\alias{get_timeseries_list}
\title{Get a list of timeseries available at gauging stations}
\usage{
get_timeseries_list(
  station_no = NULL,
  parametertype_name = NULL,
  ts_id = NULL,
  ts_name = NULL,
  ...,
  returnfields = c("station_no", "station_name", "parametertype_name", "ts_id",
    "ts_name")
)
}
\arguments{
\item{station_no}{Optionally, a numeric or character vector of station
numbers.}

\item{parametertype_name}{Optionally, a character vector of parameters. This
can be useful if, for instance, you are only interested in stations which
measure "Water Course Discharge". Use \code{\link[=list_parameters]{list_parameters()}} to see available
options.}

\item{ts_id}{Optionally, a specific time series. The id specifies both the
time series product and the station}

\item{ts_name}{Optionally, the name of a specific time series}

\item{...}{Optional named query fields which can be used to narrow the
request. Vectors of length greater than one will be collapsed into a comma
separated character string. Use \code{\link[=list_query_fields]{list_query_fields()}} to see available
options, including fields which accept a comma separated list or wildcard.}

\item{returnfields}{A character vector of columns to include in the returned
tibble. If not defined, the default columns will be determined by the API.
Use \code{\link[=list_return_fields]{list_return_fields()}} to see available options.}
}
\value{
A tibble with columns matching \code{returnfields}.
}
\description{
\code{get_timeseries_list} queries the BOM API and returns a table of timeseries
which can be downloaded for one or more gauging stations. The query can be
narrowed to specific timeseries, parameters, or stations. If no query filters
are specified the API will return an error.
}
\examples{
get_timeseries_list(station_no = 609017)
}
