% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_helpers.R
\name{list_requests}
\alias{list_requests}
\alias{list_query_fields}
\alias{list_return_fields}
\alias{list_optional_fields}
\alias{list_formats}
\alias{list_parameters}
\title{List request info}
\usage{
list_requests()

list_query_fields(request = NULL)

list_return_fields(request = NULL)

list_optional_fields(request = NULL)

list_formats(request = NULL)

list_parameters()
}
\arguments{
\item{request}{Optional request type to filter the response}
}
\value{
A tibble
}
\description{
List request info
}
\examples{
list_requests()

list_query_fields("getStationList")
}
