% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterbom-package.R
\docType{package}
\name{waterbom-package}
\alias{waterbom}
\alias{waterbom-package}
\title{waterbom: Download water data from the Australian Bureau of Meteorology}
\description{
An set of functions for accessing the data compiled by the Australian Bureau of Meteorology.
}
\author{
\strong{Maintainer}: Callum Savage \email{callum.s.savage@gmail.com}

}
\keyword{internal}
